[superproject] Roll fuchsia [virtualization] Parameterize net device tests
This runs the tests against the C++ and Rust components. This skips the
RX test for now, as the device doesn't yet support that.

This also does a cleaner shutdown (which the C++ device doesn't seem to
do). This stops some asserts firing when the component terminates, which
only happens for the Rust device because we support stopping the
component on error.

Original-Bug: 95485
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743964
Original-Revision: 1bfa5738ae63205d0de3322369c59484a74c22c0
GitOrigin-RevId: 13916c39075114c81912be6f1a087bd48d2b7318
Roller-URL: https://ci.chromium.org/b/8799438882802992673
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I8efb92bb8c6f9cce33e990236059a679fe799e7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746042
Commit-Queue: GI Roller <global-integration-roller@fuchsia-infra.iam.gserviceaccount.com>