[netstack3] Avoid panic in TCP
When binding a TCP socket, we used to first remove the unbound socket
and then add a bound entry in the socketmap. However it is hard to hold
in a later change, we added a possible early return after removing the
unbound entry and adding the bound entry. This caused panics in syscall
tests. Now it is rewritten so we only remove after we know the new entry
is successfully added. Fixing this issue exposed another potential panic
when we didn't cancel a timer.

Test: added a unit test; manually ran syscall tests locally that are not
run in CQ yet, verified that none of them will panic the netstack3 now.

Change-Id: I3f72420b311cc1d990737647687f127b067c709f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745358
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>