[devices] Pass MUST_ISOLATE flag from driver host to driver manager
Previously, the device_add API in the driver host accepted
DEVICE_ADD_MUST_ISOLATE as one of its flags, but it did not pass it
along to the driver manager. Instead, the driver manager set its own
DEV_CTX_MUST_ISOLATE flag based on some other properties of the request,
such as whether or not the device had an outgoing directory. This
meant that setting DEVICE_ADD_MUST_ISOLATE did not always produce the
intended result.

One instance of DEVICE_ADD_MUST_ISOLATE which was previously ignored,
the Banjo PCI node, is removed to get a test that uses IsolatedDevmgr
to pass. This should have no effect on actual devices because nothing
binds to the Banjo PCI node anymore after the FIDL migration.

An integration test is added to verify that MUST_ISOLATE is passed from
the driver host to the driver manager.

Fixed: 112652
Test: driver-manager-test and fidl-protocol-test
Change-Id: I5d640c0be4557149ee6863e87d630dbb468c3c41
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742905
Commit-Queue: Ian Fisher <iafisher@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>