[nxpfmac] Fix deadlock in EAPOL response handling
There was an issue with EAPOL response handling where it synchronously
acquired a lock that could be held by another call from SME. At the same
time the EAPOL response code would block the driver from servicing the
SME request. This resulted in a deadlock.

The code already performed the EAPOL response call to SME in a separate
thread. This change makes sure that the validity check of the SME
interface along with holding the lock required for the check also
happens in the separate thread.

Test: WlanSecurityComplianceABGTest ACTS test
Change-Id: I0278367dab79dca9e16adfc12e6b172dc43ffd5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745985
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>