[dhcpv6] Don't hold address hints forever
This change only holds a single address per IA_NA entry in the following
priority order:
  1) The assigned address
  2) The advertised address
  3) The preferred/configured/hint address

Previously, the DHCPv6 state machine held on to IA_NA hints for the
lifetime of the state machine. However, in practice, we only need
to hold the hints until the IA is assigned or advertised an address
and then the hint address is essentially useless.

When an address different from the hint is advertised (in response
to Solicit messages), the server is explicitly choosing to ignore
our hint. In following Request messages, the state machine uses
the advertised address and will never look at the hint address for
future requests.

Once an address is assigned, the state machine will attempt to
keep the assigned address for the lifetime of the state machine
through renew/rebind messages and will never attempt to assign
the hint address in its place.

Basically, the configured/preferred/hint address should only be held
until the server advertises or assigns us an address per its policy.

Bug: 80595
Change-Id: I2625350c6b96555164ad38b625bc973b6e663f34
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745472
Reviewed-by: Tony Gong <gongt@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>