[dhcpv6] Don't hold advertisement after assignment
As part of DHCPv6 server discovery/selection, the state machine collects
advertisements sent by servers in response to solicit messages the client
sends. These advertisements hold a candidate Reply to a future Request
message with the same options that were included in the Solicit message.

Once a set of advertisements is collected, the state machine transitions
to the Requesting state where we attempt to request and have a lease
assigned to us by the server. If this fails, the state machine will try
the next server in the set of advertisements it collected. If assignment
succeeds, the client transitions to the Assigned state and uses the
selected server for future messages (until Rebind).

If at any later point the clients needs to select a new server, the
current implementation looks at the next server in the set of
collected advertisements and tries that server. If we transitioned
into the Assigned state, we would not attempt to communicate with
a server for some time before an error occurs that requires the client
to pick an alternate server. In this time, the set of advertisements
will likely have gone stale as the server may have assigned advertised
IAs to some other client.

To avoid picking servers using stale advertisements, this change updates
the state machine to only hold onto collected advertisements in the
Server Discovery/Selection state and in the Requesting state when
transitioning from Server Discovery/Selection. When an alternate server
needs to be selected after a lease has been assigned, the state machine
will restart server discovery from scratch to make sure a fresh set of
advertisements is collected.

Bug: 80595
Change-Id: Ie27b82bb5e64065e20fca5f4dab05319ff1f6367
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745473
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Tony Gong <gongt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>