[nxpfmac] Properly support WEP
The driver had an initial attempt at supporting WEP but this was
insufficient to make it work.

This change adds the additional call needed to enable the WEP key when
configured as well as removing incorrect IE configuration and handling
the detection of shared or open WEP keys.

BUG: 108742
Test: WlanSecurityComplianceABGTest ACTS test
Change-Id: Icec0f49b62566db34761d5952846129dc0bdd8f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745986
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>