[docs] fidl-ir-path needs to be bazel-out in sdk-integration
Bazel SDK now performs toolchain transitions, which switches the
toolchain during a build command. This causes bazel-bin to end up
symlinking to the output of the last toolchain used in the build
command, and not necessarily the toolchain where the full fidl IR
was built. bazel-out has all the output directories but is
considerably larger and slower, but at least fidlcat works now.

Fixed: 112860
Change-Id: Ib7cae454ca7fd1eb237665c8c5eba59fac841961
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746303
Commit-Queue: Wayne Piekarski <waynepie@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>