[line_input] Keep raw input mode when prompt is hidden.
Previously the line_input would revert the terminal mode back to the
original one when it is hidden to allow output to be printed like
normal. But this also put back normal input mode so control-C
would exit the program (rather than call the associated callback) and
typing would be echoed to the terminal, even without being processed.

This change creates a new "hidden" terminal mode. This mode keeps the
raw input flags so input is processed consistently. It now also
explicitly sets the output flags we expect for "normal" output so it
will be correct even if the terminal was set to a weird mode when the
process starts.

Bug: 112841
Change-Id: I25f81078627b9f993cc4ea68a0399ce6630af0b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745347
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>