[debugger] Add thread controller completion CBs.
Thread controllers have a callback on init so that the caller can know
if the controller initialized itself properly (useful when there are
requirements about breakpoints needing to be set).

But there was no way to know when a controller thinks it's "done." This
adds an optional constructor callback so the creater of the controller
can observe this event.

This callback is currently unused and there should be no behavior
change. It will be used in a follow-up CL.

Change-Id: I5271d7f02ff61dc4e7953bf098b7d662570511b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744786
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>