[netstack] Disable NS3 ifwatcher test variant
The NS3 variant of `test_watcher` was accidentally running with NS2.

The test does not trivially pass against NS3 due to differences
between the two netstacks. A bug has been filed to split this test up
so it is easier to enable against NS3 and account for differences
between the Netstacks and a TODO added to address this.

Bug: 112627
Change-Id: I60eb09ddf56d06dc21504e055a7c7d18913f6031
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745088
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>