[superproject] Roll llvm-project/libc [libc] add performance options for string to float
This allows the client to set compile flags to disable the passes that
the string to float function uses. A client may be willing to trade off
performance for a reduction in code size, and this allows for that
fine-tuning.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D136495

GitOrigin-RevId: a874bab97c09996d38d304a91085a824cac04baa
Original-Revision: 31449b5cf19c9f3152026abbdb1836345e1a0090
Roller-URL: https://ci.chromium.org/b/8799386778149413361
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I3a3a0a33b8406b3ae620a8ba5b653ae6d1c43bbb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744571