[flatland] Protected render targets in DisplayController
This CL adds protected render targets to Flatland so that we can switch
to them if the render has any images that are backed by protected
memory.

This CL also adds an integration test that exercises this codepath on
devices that support protected memory.

Bug: 73662
Change-Id: I023b84828a59f881b2fc9756ea8d996f5fd3316a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734682
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Emircan Uysaler <emircan@google.com>