[test_manager] Don't crash if logger connection fails
Return an error from the hermetic resolver instead of crashing the
entire process when connecting to LogSink fails. This connection error
is expected in cases where the hermetic resolver is running before
Archivist but the test realm is being shut down. Because hermetic
resolver is a mock and runs as a Rust task on an executor, it could run
for a little while even after stop is called. As a result, it may
attempt to connect to LogSink in this situation, and the connection
could fail.

Bug: 112984
Multiply: run_test_suite_integration_tests, test_manager_test

Change-Id: I089ef241f00a0555084df0b0752be2028f517329
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746322
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>