[ffx][test] Update help output
Update the help output for ffx test run in preparation for running it
through a design review. Reordering the options within the struct is
relevant as the order in the source code determines the order in which
the options are printed in the help message.

Change-Id: Ifaf341e088f266ca956e64c95cbaf5b2b936de29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743082
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>