[display] ClientProxy is no longer a device
Make ClientProxy strictly a FIDL server instead of an instance
device.

This means that Controller needs to manage the lifetime of the
ClientProxies. It does this by keeping a set of them, and freeing
them when the client end is closed. The only place a ClientProxy
will be freed is in the unbind callback of its FIDL server
when it calls OnClientDead.

This removes the use of DEVICE_ADD_INSTANCE which is being
deprecated.

The device_channel still has to live on the ClientProxy but it
will be removed in a followup CL.

CloseOnControllerLoop is updated to not wait for the ClientProxy
to finish closing. This is fine, because if it does not finish
in time we will be calling loop_.Shutdown() and freeing all of
the Clients anyway.

Change-Id: I62ad493322b9d8188111929217844fa234a1e3d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745149
Reviewed-by: John Bauman <jbauman@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>