[display] Remove unnecesary device controller
This controller handle was being passed to the DDK function
add_device_controller, which the Driver Framework is trying to remove.

It looks like the only use of this handle was to keep the device alive.
If this handle is never passed in the first place, the device will be
kept alive by the controller handle.

Most of the comments in the removed code suggest that this device handle
is entirely unused.

Fixed: 33675

Change-Id: I2a8048a999add225da03a27b5d99c47511d41317
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744863
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
API-Review: John Bauman <jbauman@google.com>