[fidl][cpp] Synchronized ServerBinding
Adds {fidl,fdf}::ServerBinding, a thread-unsafe server binding object
that trades threading restrictions for synchronous unbinding.

fidl::ServerBinding dispatches messages from a Zircon channel.
fdf::ServerBinding dispatches messages from a driver channel. They share
most of the implementation but the comments are duplicated for ease of
user learning.

The API is roughly modeled off of fidl::Binding. |Unbind| was omitted
because it's less often used, and is not trivial to add: it would need
to extract the channel from the wrapped ServerBindingRef. We could add
it in a backwards-compatible way in the future.

fidl::CloseHandler is not exactly equivalent to the set_error_handler in
HLCPP: the former would surface both user-initiated and external error
teardown, whereas the latter only surfaces error. This new behavior
proved to be more robust because people don't have to duplicate cleanup
logic in both the user-initiated teardown code paths and error code paths.
In the new C++ bindings, there are a lot more ways to user-initiate a
teardown: every completer has a .Close() function. Now all these would
feed into a CloseHandler callback which is a central place to perform
other cleanup.

The tests are also modeled off of
//sdk/ctf/tests/pkg/fidl/cpp/binding_unittest.cc. Added new tests that
test we reliably crash if the user tried to use the binding from
multiple threads or from an incorrect driver dispatcher.

I migrated some examples to fidl::ServerBinding. The examples are all
"self-managed" use cases so they don't really show an idiomatic use of
fidl::ServerBinding -- a binding set is arguably a better fit.
fidl::ServerBinding is useful during simple tests, or servers with only
a single client connection (very rare), or when the user needs a custom
data structure to hold the servers where binding set doesn't meet the
needs. E.g. in flatland it's common to maintain a ID-to-server
dictionary and have servers managed in that dictionary.

Bug: 95685

Change-Id: I24ea73399b4b9e3a0e310d1b7a841e611fc266fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743282
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>