[sdk/ctf] Add support for CTF test realms
This change allows CTF test authors to create test realms
tailored for running specific CTF tests. The test realms
are made available at build time as subpackages to all
ctf_fuchsia_test_package targets that specify
'enable_ctf_test_realms = true'. After a trial period this
will become the default.

A CTF test realm contains a test's dependencies and is not
bundled together with the test when it is released. Instead
the realms are always built at HEAD and provided to the test.

As an example, the fuchsia.driver.test CTF test is rewritten.

Change-Id: I6745b5f778af744abf10f3235404d588d1d7e3b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735588
Reviewed-by: Victor Chang <vfcc@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Kendal Harland <kjharland@google.com>