[kernel][hypervisor] Avoid permission faults
The hpfar_el2 register, which is relied upon to give the IPA to resolve
a data or instruction abort, is not valid for permission faults.
Permission faults therefore cannot be handled, and must be avoided.

This change modifies the protect path for stage 2 aspaces to perform an
unmap instead, and adds checks to the instruction and data abort
handlers to ensure the hpfar_el2 is valid before attempting to use it.

Change-Id: I5ff6e069b4c2ea92fa8f028a4364ece01661c367
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744825
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>