[ktrace] Migrate VCPU Events to FXT
Because this means we no longer do a post processing step, we don't know
the exit type of the event when we emit the vcpu begin event. To work
around this, the exit type has been moved to become an argument and the
even name is now just "vcpu".

Bug: 98176
Change-Id: I1ee6aef91b33ea702c4acdabc80392f695e812fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743067
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>