[vulkan] Wait for old command buffers before importing semaphores.
New versions of the Validation layers give this warning in
SwapchainFidlTestSuite/SwapchainFidlTest.PresentAndAcquireNoSemaphore/UnprotectedNoCopyValidationAfter:

Cannot call vkImportSemaphoreZirconHandleFUCHSIA on VkSemaphore
0xa000000000a[] that is currently in use by a command buffer.

This is because we wait in acquire using the zircon semaphore, so they
don't realize that the command buffer must necessarily be completed.
We can signal a fence in the present submit and wait on that before
doing vkImportSemaphoreZirconHandleFUCHSIA to ensure that the validation
layers are aware of the behavior.

Bug: 112606
Change-Id: I3da5f7a12b9396ce61f20e097a10f1f5fd0a2ca8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745464
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: John Rosasco <rosasco@google.com>
Commit-Queue: John Bauman <jbauman@google.com>