[dev-ec] Add traces for lid close events.
Recently there have been a number of reports of devices not shutting
down when the lid is closed. The EC sees lid closed events however the
lid driver does not. This bug is difficult to reproduce so we're adding
some traces so that next time it happens the logs will be more helpful.

Bug: 111460

Change-Id: I99ab295af857a1e0f61cb3c18ee75cd5976f62a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745047
Reviewed-by: Simon Shields <simonshields@google.com>
Commit-Queue: Sarah Pham <smpham@google.com>