[vaapi] Set min_bytes_per_row for chromium
Currently the ImageFormat bytes_per_row field is not used which means
that the codec can only change the stride when new buffers are allocated
and the constraints are set. Instead of using bytes_per_row we must set
the min_bytes_per_row as a part of the output constraints and then stick
to that value so that we and chromium agree on the buffer stride.

Bug: 112011
Change-Id: I83a8c91c7ceb5b4070a757a13dfa21532fe9a35c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/746703
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>