[starnix] BinderProcess must not outlive the given process
BinderProcess was owned by the fd that created it. Because the fd can be
duplicated and passed to another process, it allowed BinderProcess to
outlive the process itself. Also, it allowed another process to access
the fd.

To prevent this, this CL changes the fd to only contains the pid of the
process. On each access the process is retrieved from the driver, and it
also checks that the pid is the one of the calling process.

Bug: 101486

Change-Id: I526cc7c4ce831274613ae67f5aa10e4f57f600f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745143
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>