[test_manager] Serve hermetic resolver even if logging fails
The hermetic resolver failing is a fatal error for tests. Currently,
connecting to the logger seems to fail flakily a few times a day. Since
this could happen for any test, we should expect to see this flake
randomly affecting all tests. This adds a fallback to not log warnings
if connecting to the logger fails instead of stopping the resolver
completely.

Bug: 112984
Multiply: run_test_suite_integration_tests

Change-Id: I915a15309d6d2c5398d3009f129e0c7ef3abf38e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747602
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>