[debugger] Don't return LLVM types from LineTable
Previously the DwarfUnit and LineTable classes returned an
llvm::DWARFDie object. This is a layering violation as we prefer to keep
the LLVM types contained in the low-level symbol factory
implementations, and there's not much one can do with this structure.

Now that the symbols expose their DIE offset and can be created from
this offset in the symbol factory, we can replace these LLVM types with
a simple symbol offset. This also lets us make some functions in the
DwarfSymbolFactory private.

This lets us clean up some references to DwarfSymbolFactory and replace
with the abstract SymbolFactory base class.

There should be no behavior change.

Change-Id: Ie8e4d15ca1668277f132c48efc6c70c5eed5bc66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744576
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>