[elfldltl] Add filesz() to ZeroFillSegment
This makes it much easier to get the filesz from a segment variant
by using a visitor like `[](auto &&s) {/*something with s.filesz()*/}`

The alternative would be to have an overloaded operator() for each
segment type, which isn't that ergonomic for something that should
be simple. The cost of this method is also basically 0.

Bug: 68283
Change-Id: I8982199b3a27eb8ace3a354d89eb126677e08875
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747036
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Brachet <abrachet@google.com>