[netstack3] Make TCP `handle_timer` more strict
Currently, `handler_timer` is lax about if the connection ID given is
valid or not and only do work if the ID is valid. This has papered over
some other bugs. This CL strengthens the precondition of the function
so that we can catch those earlier.

Change-Id: I7fbde88778d121fc4d26f13b70f9742ee5b502d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/745523
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Alex Konradi <akonradi@google.com>