[vaapi] Reset decoder on stream stop
If the stream is stopped then the decoder can reset its context. For
tiled decoder which DPB are sysmem buffers, this matters because
CodecImpl can call CoreCodecEnsureBuffersNotConfigured() after the
stream is stopped and if the decoder's context is not reset it will
still have reference frames that are backed by sysmem output buffers.
This can cause ASSERTs to fail since it is expected that all buffers can
be deconfigured after a stream is stopped.

Fix: 110813
Bug: 111139
Change-Id: Ib5633ac1c97e7b65d3281e83a9a4f74752c5d410
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739584
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>