[build] Stop publishing <hash>.root.json
We will be migrating to a rust-tuf based system for creating TUF
metadata, and so we'll no longer produce the TUF-0.9-era
<hash>.root.json file. We never used this file on a production system,
so it should be safe to remove this file.

This also changes the system to copy the `root.json` from the build,
rather than inferring it from one of the root metadata files.

Change-Id: Idd69883aa2f05c02c67d4c1e07371000f79414d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/747951
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>